*: add support for rootless containers
This enables the support for the rootless container mode. There are
certain restrictions on what non-root users can do, resulting in several
runC features not being used. There are no checks in place at the moment
to make this clear to users.

* All cgroup operations require having CAP_SYS_ADMIN in the root user
namespace. This means that we cannot set up *any* cgroups, or join
cgroups. This can be circumvented by having the user own the current
cgroup [this mode is currently not implemented].

* setgroups(2) cannot be used in a non-privileged user namespace setup,
so we have to hard fail if for some reason we decide that we need
setgroups(2) enabled.

* We cannot map any user other than ourselves in a rootless container,
which means that any user-related directives won't work. You can only be
"root".

Signed-off-by: Aleksa Sarai <asarai@suse.de>